# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CartFlow API",
    "version": "1.0.0",
    "description": "Production-ready multi-tenant e-commerce cart engine API."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "headers": {
      "XRequestID": {
        "description": "Unique request tracking identifier for observability.",
        "schema": {
          "type": "string",
          "example": "req_123456"
        }
      },
      "IdempotencyKey": {
        "description": "Unique key used to prevent duplicate request processing.",
        "schema": {
          "type": "string",
          "example": "item-add-123"
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Error message details" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": { "type": "string", "example": "Validation failed" },
          "details": {
            "type": "array",
            "items": { "type": "object" }
          }
        }
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "productId": { "type": "string" },
          "productName": { "type": "string" },
          "quantity": { "type": "number" },
          "unitPrice": { "type": "number" },
          "category": { "type": "string" }
        }
      },
      "Cart": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "status": { "type": "string", "enum": ["ACTIVE", "EXPIRED", "CHECKED_OUT"] },
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CartItem" }
          },
          "subtotal": { "type": "number" },
          "discount": { "type": "number" },
          "total": { "type": "number" },
          "version": { "type": "number", "description": "Optimistic Concurrency Control version tracking" }
        }
      }
    }
  },
  "tags": [
    { "name": "Auth", "description": "Authentication endpoints" },
    { "name": "Cart", "description": "Cart management operations" },
    { "name": "Checkout", "description": "Checkout and promotion calculation" },
    { "name": "Analytics", "description": "System analytics and aggregates" }
  ],
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "securepassword123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "headers": { "X-Request-ID": { "$ref": "#/components/headers/XRequestID" } },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string" },
                            "email": { "type": "string" }
                          }
                        },
                        "token": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation Error or User Already Exists", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "429": { "description": "Too Many Requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate user and get JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "user@example.com" },
                  "password": { "type": "string", "example": "securepassword123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": { "X-Request-ID": { "$ref": "#/components/headers/XRequestID" } },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "Too Many Requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/cart": {
      "get": {
        "tags": ["Cart"],
        "summary": "Get active cart for the authenticated user",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Cart retrieved successfully",
            "headers": { "X-Request-ID": { "$ref": "#/components/headers/XRequestID" } },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Cart" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "Too Many Requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/cart/items": {
      "post": {
        "tags": ["Cart"],
        "summary": "Add or update an item in the cart. Send quantity 0 to remove.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "example": "item-add-12345",
            "description": "Unique key used to prevent duplicate request processing."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": { "type": "string", "example": "prod_123" },
                  "productName": { "type": "string", "example": "Wireless Mouse" },
                  "quantity": { "type": "number", "example": 2, "description": "Quantity to add. Set to 0 to remove." },
                  "unitPrice": { "type": "number", "example": 25.5 },
                  "category": { "type": "string", "example": "electronics" }
                },
                "required": ["productId", "productName", "quantity", "unitPrice", "category"]
              },
              "example": {
                "productId": "prod_123",
                "productName": "Wireless Mouse",
                "quantity": 2,
                "unitPrice": 25.5,
                "category": "electronics"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart updated successfully",
            "headers": { "X-Request-ID": { "$ref": "#/components/headers/XRequestID" } },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Cart" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Idempotency Conflict or Cart modified by another request (OCC)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "Too Many Requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/checkout": {
      "post": {
        "tags": ["Checkout"],
        "summary": "Process checkout and calculate final promotional discounts",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": { "type": "string" },
            "example": "checkout-tx-789",
            "description": "Unique key used to prevent double charging or processing."
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cartId": { "type": "string", "example": "60a1b2c3d4e5f60012345678" }
                }
              },
              "example": {
                "cartId": "60a1b2c3d4e5f60012345678"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout processed successfully",
            "headers": { "X-Request-ID": { "$ref": "#/components/headers/XRequestID" } },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "subtotal": { "type": "number", "example": 52000 },
                    "discounts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "campaignName": { "type": "string", "example": "Tier 3 Cart Discount" },
                          "amount": { "type": "number", "example": 10400 }
                        }
                      }
                    },
                    "finalAmount": { "type": "number", "example": 41600 },
                    "appliedCampaigns": {
                      "type": "array",
                      "items": { "type": "string", "example": "Tier 3 Cart Discount" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Cannot checkout an empty cart", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "409": { "description": "Idempotency Conflict or Cart modified by another request (OCC)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "Too Many Requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/v1/analytics/dashboard": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Retrieve aggregated system analytics dashboard",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Analytics retrieved successfully",
            "headers": { "X-Request-ID": { "$ref": "#/components/headers/XRequestID" } },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "averageCartValue": { "type": "number", "example": 4800 },
                        "totalCheckouts": { "type": "number", "example": 156 },
                        "abandonedCarts": { "type": "number", "example": 24 },
                        "topProducts": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "productId": { "type": "string" },
                              "count": { "type": "number" }
                            }
                          }
                        },
                        "topCategories": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "category": { "type": "string" },
                              "count": { "type": "number" }
                            }
                          }
                        },
                        "mostUsedCampaigns": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "campaignId": { "type": "string" },
                              "count": { "type": "number" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "Too Many Requests", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal Server Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    }
  }
}